requirements.txt
opencv-python-4.0.0.21
numpy-1.16.1
mss-4.0.1
pyautogui-0.9.41
PyUserInput-0.1.11
